Implement driver for HTS221 temperature sensor
The HTS221 is a temperature and humidity sensor present on the Nano BLE
33 Sense board.

The HTS221 requires adjusting readings based on per-chip calibration
registers in the chip. There are two main downsides of the
implementation in this commit. First, the calibration registers are read
on every sensor reading instead of reading them once either lazily or
upon board initialization. Second, adjusting readings requires
computing a simple linear regression over two data points in the
callibration registers to find a slope and intercept---the current
implementation uses floating point computations to perform the
regression, which may have high overhead on platforms without hardware
floating point.